refactor(reminders): rename AlarmManagerService to ReviewReminderAlarmManager
Now that snoozing and the shared alarm helpers live elsewhere, what is left is
neither a service nor generic: it only schedules review reminder notifications.
Move it to the reviewreminders package and name it accordingly.
